+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Change Log
 
-
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
+
+## [0.2] - 2019-04-23
+### Changed
+- `JSONAutoTest`: added more tests (derived classes)
+- `JSONAuto`: added code to check for `@JSONName` annotation
+- `JSONAutoTest`: added tests for `@JSONName` annotation
+
+### Added
+- `JSONName`: added annotation
+
 
 ## [0.1] - 2019-04-22
 ### Added